3 listeners attribute to api
